requirements.txt
scapy-python3==0.20
click==6.7
cloudpickle==0.3.1
dask==0.15.0
distributed==1.17.1
HeapDict==1.0.0
msgpack-python==0.4.8
psutil==5.2.2
six==1.10.0
sortedcontainers==1.5.7
tblib==1.3.2
toolz==0.8.2
tornado==4.5.1
zict==0.1.2

